PDFBOX-2262: Increase width precision
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/branches/no-awt@1620486 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/pdmodel/font/PDTrueTypeFont.java
@@ -141,8 +141,8 @@ public class PDTrueTypeFont extends PDSimpleFont
     protected float getWidthFromFont(int code) throws IOException
     {
         int gid = codeToGID(code);
-        int width = ttf.getAdvanceWidth(gid);
-        int unitsPerEM = ttf.getUnitsPerEm();
+        float width = ttf.getAdvanceWidth(gid);
+        float unitsPerEM = ttf.getUnitsPerEm();
         if (unitsPerEM != 1000)
         {
             width *= 1000f / unitsPerEM;
